backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.6.0
python-dotenv==1.0.0
psycopg2-binary
gunicorn==21.2.0
typing-extensions==4.10.0
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
